+++ src/main/java/lumi/insert/app/core/repository/projection/ProductRefreshProjection.java
@@ -1,7 +1,9 @@
 package lumi.insert.app.core.repository.projection;
 
+import java.math.BigDecimal;
+
 public record ProductRefreshProjection(
     Long id, 
-    Long sellPrice, 
-    Long stockQuantity) {
+    BigDecimal sellPrice, 
+    BigDecimal stockQuantity) {
 }
