+++ src/main/java/lumi/insert/app/dto/request/TransactionPaymentCreateRequest.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 import org.springframework.web.multipart.MultipartFile;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -25,7 +27,7 @@ public class TransactionPaymentCreateRequest {
     @NotNull(message = "totalPayment must not be null")
     @Min(value = 1, message = "totalPayment minimal value is 1")
     @Schema(description = "Payment amount", example = "100000")
-    Long totalPayment;
+    BigDecimal totalPayment;
 
     @Schema(description = "File(image) to be store")
     @NotNull(message = "files must not be null")
