+++ src/main/java/lumi/insert/app/dto/response/ProductStockResponse.java
@@ -1,5 +1,7 @@
 package lumi.insert.app.dto.response;
 
+import java.math.BigDecimal;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
@@ -11,7 +13,7 @@ public record ProductStockResponse (
     Long id, 
     
     @Schema(description = "The current number of units available in the warehouse/store", example = "42")
-    Long stockQuantity
+    BigDecimal stockQuantity
     
 )  implements Identifiable {
